requirements.txt
Click==7.0
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.2
opencv-python==4.1.1.26
pycodestyle==2.5.0
pyflakes==2.1.1
pyserial==3.4
Werkzeug==0.16.0
rubik-solver==0.2.0
